fix under the norm
