+++ inc/cmd.h
@@ -6,7 +6,7 @@
 /*   By: dayano <dayano@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/13 13:47:36 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/04/29 16:09:43 by dayano           ###   ########.fr       */
+/*   Updated: 2025/04/29 16:53:33 by dayano           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,7 @@
 # define CMD_H
 
 # include <unistd.h>
+
 typedef enum e_redir_type
 {
 	REDIR_NONE,
