+++ src/invoke_cmd/exec_pipeline.c
@@ -6,48 +6,13 @@
 /*   By: dayano <dayano@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/21 13:14:15 by dayano            #+#    #+#             */
-/*   Updated: 2025/04/29 16:36:09 by dayano           ###   ########.fr       */
+/*   Updated: 2025/04/29 16:52:01 by dayano           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "main.h"
 
-void	init_fds(int fds[2])
-{
-	fds[0] = -1;
-	fds[1] = -1;
-}
-
-void	cleanup_fds(int fds1[2])
-{
-	if (fds1[0] != -1)
-		close(fds1[0]);
-	if (fds1[1] != -1)
-		close(fds1[1]);
-}
-
-void	swap_fds(int prev_fds[2], int curr_fds[2])
-{
-	prev_fds[0] = curr_fds[0];
-	prev_fds[1] = curr_fds[1];
-}
-
-void	create_pipe(bool need, int fds[2], t_cmd *cmd)
-{
-	if (!need)
-	{
-		fds[0] = -1;
-		fds[1] = -1;
-		return ;
-	}
-	if (pipe(fds) < 0)
-	{
-		print_error(cmd->argv[0]);
-		exit(EX_OSERR);
-	}
-}
-
-void	setup_stdin(t_cmd *cmd, t_cmd *cmd_head, int prev_fds[2])
+static void	_setup_stdin(t_cmd *cmd, t_cmd *cmd_head, int prev_fds[2])
 {
 	if (is_head(cmd, cmd_head) || prev_fds[0] < 0)
 		return ;
@@ -60,7 +25,7 @@ void	setup_stdin(t_cmd *cmd, t_cmd *cmd_head, int prev_fds[2])
 	close(prev_fds[1]);
 }
 
-void	setup_stdout(t_cmd *cmd, int curr_fds[2])
+static void	_setup_stdout(t_cmd *cmd, int curr_fds[2])
 {
 	if (!is_tail(cmd) || curr_fds[1] >= 0)
 	{
@@ -74,7 +39,7 @@ void	setup_stdout(t_cmd *cmd, int curr_fds[2])
 	}
 }
 
-pid_t	fork_command(t_cmd *cmd, t_cmd *cmd_head, t_pipe_io *pipefds,
+static pid_t	_fork_command(t_cmd *cmd, t_cmd *cmd_head, t_pipe_io *pipefds,
 		t_minish *minish)
 {
 	pid_t	pid;
@@ -87,8 +52,8 @@ pid_t	fork_command(t_cmd *cmd, t_cmd *cmd_head, t_pipe_io *pipefds,
 	}
 	if (pid > 0)
 		return (pid);
-	setup_stdin(cmd, cmd_head, pipefds->prev_fds);
-	setup_stdout(cmd, pipefds->curr_fds);
+	_setup_stdin(cmd, cmd_head, pipefds->prev_fds);
+	_setup_stdout(cmd, pipefds->curr_fds);
 	if ((cmd->next != NULL) && is_redirect(cmd->next))
 		redirect(cmd->next);
 	if (is_builtin(cmd))
@@ -99,10 +64,13 @@ pid_t	fork_command(t_cmd *cmd, t_cmd *cmd_head, t_pipe_io *pipefds,
 }
 
 /**
- * @brief
+ * @brief Execute a pipeline of commands.
  *
- * @param cmd_head
- * @param minish
+ * For each command, create a pipe if needed, fork a child, redirect its
+ * stdin/stdout through the pipes, and record its PID for waiting.
+ *
+ * @param cmd_head  First command in the pipeline.
+ * @param minish    Shell context for built-ins and execution.
  */
 void	exec_pipeline(t_cmd *cmd_head, t_minish *minish)
 {
@@ -117,70 +85,10 @@ void	exec_pipeline(t_cmd *cmd_head, t_minish *minish)
 	{
 		need_pipe = !is_tail(cmd);
 		create_pipe(need_pipe, pipefds.curr_fds, cmd);
-		cmd->pid = fork_command(cmd, cmd_head, &pipefds, minish);
+		cmd->pid = _fork_command(cmd, cmd_head, &pipefds, minish);
 		cleanup_fds(pipefds.prev_fds);
 		swap_fds(pipefds.prev_fds, pipefds.curr_fds);
 		init_fds(pipefds.curr_fds);
 		cmd = cmd->next;
 	}
-}
-
-/**
- * @brief Wait for one command’s process to finish.
- *
- * This function calls waitpid() for cmd->pid, stores the exit
- * or signal status into cmd->status, and returns that status.
- *
- * @param cmd  The command whose child process to wait for.
- * @return     The exit status or 128+signal number on signal,
-	or EXIT_FAILURE on error.
- */
-static int	_get_status(t_cmd *cmd)
-{
-	int	ret;
-	int	status;
-
-	ret = 0;
-	ret = waitpid(cmd->pid, &status, 0);
-	if (ret < 0)
-	{
-		cmd->status = EXIT_FAILURE;
-		return (cmd->status);
-	}
-	if (WIFEXITED(status))
-		cmd->status = WEXITSTATUS(status);
-	else if (WIFSIGNALED(status))
-		cmd->status = 128 + WTERMSIG(status);
-	else
-		cmd->status = EXIT_FAILURE;
-	return (cmd->status);
-}
-
-/**
- * @brief Wait for all child processes in the pipeline to finish.
- *
- * This function walks through each command in the pipeline,
- * calls waitpid() for every non-redirect node, stores each
- * child’s exit or signal status into cmd->status, and then
- * returns the status of the last command executed.
- *
- * @param cmd_head  Pointer to the head of the command list.
- * @return          The exit status of the last command in the pipeline.
- */
-int	wait_pipeline(t_cmd *cmd_head)
-{
-	int		last_status;
-	int		ret;
-	t_cmd	*cmd;
-
-	last_status = 0;
-	ret = 0;
-	cmd = cmd_head;
-	while (cmd)
-	{
-		if (!is_redirect(cmd))
-			last_status = _get_status(cmd);
-		cmd = cmd->next;
-	}
-	return (last_status);
 }
